Make initialiser object a pointer
In the simulator class. This pointer is kept so that the attributes
of the object it points to can be used (for example, the random number
generators that C++11 provide in its <random> header).

Fixed generation of position in rectangular_source class (use two random
values instead of only one...)