Major update: 2x speed up
The speed up has been achieved at the cost of 'correctness' in terms of
the code written: a lots of #defines have been introduced for the
necessary operations with vectors.

These #define assume the parameters to be vec3, a new type whose name is
borrowed from the glm library, and implement operations between them,
like addition, dot product, ... Find these defines in the math/ directory.

In order to avoid conflicts these #defines start with '__pm'
(physics::math), since they are defined inside the namespace
physim::math (although it is not needed to indicate the namespaces when
using these defines).

The type vec3 is a struct with three members __x,__y,__z with operators
for easy, but not as efficient as possible, manipulation.

In the future, these defines may be removed.

In the process improving the efficiency some methods were removed.
However, these were not important. For example, particles can no longer
be deleted from the simulator.