Major update in examples
What used to be called 'render' is now called base. It includes several
new files: types of cameras, and a box structure.

The simple viewer in opengl-qt has been removed and replaced by an
opengl-only viewer (that uses glut).

It is planned that in the near future the Qt-based particles renderer
will be replaced by another opengl-only renderer.