Use shared pointers for model compilation
So far, whenever a model (for rendering sphere, in this case) had to be
displayed, for the sake of efficiency it was compiled into a GL list
(I know, there are more efficient ways of rendering something...). Also,
a model had to be assigned to a rgeom object.

But this object would free its model in its destructor, so each model
had to be (hard) copied for every rgeom object (otherwise, there would
be double frees).

To avoid (hard) copying an object several times (and also compiling it
several times), shared pointers are now being used.