Added image loading functions
Two files added: the first is used only to include the second.
The second is a file that is part of a well-know C++ header-only
library. The library can be downloaded from the repository

https://github.com/nothings/stb

This file was downloaded and copied 'as is' from this repository.
No changes were made to the file.