Several changes
Indices read in a .obj file are in the range [1,num vertices]. Normalise
them to the range [0,num vertices).

Geometrical object stores the triangles directly, not the vertices and
then the triangles as indices. This is done to speed up the future
intersection tests (there will not be need to construct an object for
the triangles each time we perform the intersection tests).