One main change. Other minor changes
Renamed namespace 'geometry' to geometric. I really made a mistake
naming a namespace with the same name of a class.

Added method to class octree to retrieve the boxes of the cells.

Render the boxes of the objects in 'particles' simulations.
Render the boxes of the objects in 'spring-meshes' simulations.
Updated command line with changes in library (renaming of namespace).